+++ generator/src/string/classic/padStart.ts
@@ -1,0 +1,4 @@
+const padStart =
+  (maxLength: number, fillString?: string) =>
+  (text: string): string =>
+    text.padStart(maxLength, fillString);
